+++ frontend/test/services/wallet.test.js
